requirements.txt
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
